requirements.txt
agno==1.6.3
google-genai==1.15.0
httpx==0.28.1
psycopg2
rich==14.0.0
fastapi==0.115.8
SQLAlchemy==2.0.41
uvicorn~=0.34.0
duckdb==1.3.0
pydantic-settings
python-dotenv
jinja2
streamlit
pypdf
lancedb
tantivy
pylance
bs4
pymongo
pymongo[srv]
fastembed
qdrant-client